Added missing Freetype-lib and added union to rect-struct.
Former-commit-id: 1bed6853301bc3f21a7816385bdcf81eebd6d2e7

+++ src/win32_main.cpp
@@ -557,7 +557,7 @@ int main(int Argc, char** Args)
         debug_info TempDebugInfo = {};
         TempDebugInfo.Header = PushString(&DebugArena, "Temporary memory");
         AddDebugValue(&DebugArena, &TempDebugInfo, "Blocks", Win32MemoryState.TempCount);
-        AddDebugValue(&DebugArena, &TempDebugInfo, "Total allocated", Win32MemoryState.TempSizeAllocated);
+        AddDebugValue(&DebugArena, &TempDebugInfo, "Total allocated", (i32)Win32MemoryState.TempSizeAllocated);
         
         auto DebugMemoryInfo = GameMemory.DebugState->DebugMemoryInfo;
         
@@ -567,14 +567,14 @@ int main(int Argc, char** Args)
         debug_info PermDebugInfo = {};
         PermDebugInfo.Header = PushString(&DebugArena, "Permanent memory");
         AddDebugValue(&DebugArena, &PermDebugInfo, "Blocks", Win32MemoryState.PermanentBlocks);
-        AddDebugValue(&DebugArena, &PermDebugInfo, "Total allocated", Win32MemoryState.PermanentSizeAllocated);
+        AddDebugValue(&DebugArena, &PermDebugInfo, "Total allocated", (i32)Win32MemoryState.PermanentSizeAllocated);
         
         GameMemory.DebugState->DebugMemoryInfo.DebugInfo[GameMemory.DebugState->DebugMemoryInfo.DebugInfoCount++] = PermDebugInfo;
         
         debug_info TotalDebugInfo = {};
         TotalDebugInfo.Header = PushString(&DebugArena, "Total memory");
         AddDebugValue(&DebugArena, &TotalDebugInfo, "Blocks", Win32MemoryState.TempCount + Win32MemoryState.PermanentBlocks);
-        AddDebugValue(&DebugArena, &TotalDebugInfo, "Total allocated", Win32MemoryState.TempSizeAllocated + Win32MemoryState.PermanentSizeAllocated);
+        AddDebugValue(&DebugArena, &TotalDebugInfo, "Total allocated", (i32)Win32MemoryState.TempSizeAllocated + (i32)Win32MemoryState.PermanentSizeAllocated);
         
         GameMemory.DebugState->DebugMemoryInfo.DebugInfo[GameMemory.DebugState->DebugMemoryInfo.DebugInfoCount++] = TotalDebugInfo;
 #endif
